requirements.txt
asgiref==3.3.4
bleach==3.3.0
Django==3.2.4
django-debug-toolbar==3.2.1
django-markdownify==0.9.0
importlib-metadata==4.5.0
Markdown==3.3.4
packaging==20.9
Pillow==8.3.1
pyparsing==2.4.7
python-environ==0.4.54
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
webencodings==0.5.1
zipp==3.4.1
